Needed files
These are all the needed files for the game the Doge cube code.txt is all the code for the game except pygame. all the code in Doge cube code is ready to paste and the Assets are all the sound effects, bg, and music. 
Little FYI the game is wrote in python 3.9.0 so keep that in mind

Youtube video on how to install pygame - https://www.youtube.com/watch?v=EKjALzLLgVs&ab_channel=ProgrammingKnowledge2